Remove puzzle descriptions and inputs
I found the GitHub repo of one of the AoC players,
<https://github.com/petertseng/adventofcode-common> where it's claimed that
"the act of publicly posting inputs is not actively discouraged, it is still
preferred that we do not do it." In addition, it is claimed that "there
definitely is an edict against posting puzzle statements."

Searching the website, I could find nothing to reference either position.
Following a that user's links took me to a couple Redit posts where those
positions are repeated, but didn't look "official." Still, I've decided that
I'll follow along and not publicly post neither the puzzle text mor the input
data.

To that end, what has been posted is being removed from this repo, and the
.gitignore is being updated to enforce that, should I forget later.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>