Add new file readers
This year looks to have a diffeent format fot the input files. Groups of
records are separated by a blank line. Records might be across lines and
lines might be groups of records. The two new functions handle, with luck,
the new formats I expect to see this year. The changes have been backported
to earlier days where appropriate.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>