Alternate solution for day 8
Wrote a new version of the solution using a module which I can use in other
challenges, if needed, this year.

Moved some modules to a new repo for all common files between AoC years.

The AoC-Common repo now holds the missing files.

Updated the main launch file to use the new module library. Added the ability
for alternate solutions to be launched by the main file.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>