Alternative for day 9
Another coder used a method for keeping a running update of the min and max for
the current list of numbers. My version used a pair of function calls against
the slice of the array, once found.

Wanted to run time trials and see how much faster it would be.

It's **not** faster. On average it clocked in at ~3.4% longer per run.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>