Clean up reporting section of programs.
The initial version of the reporting looks ok when doing a daily program.
However, when doing a run from the beginning to the end, or current day, it can
get rather verbose. Trimming it hurts nothing and saves a miniscule amount of
processor time. Not that anyone cares about that now.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>